+++ nix/workbench/analyse/analyse.sh
@@ -99,6 +99,15 @@ then backend=$WB_BACKEND
 else backend=
 fi
 
+if test -v "WB_LOCLI_DB"
+then storage=$WB_LOCLI_DB
+else storage=0
+fi
+if [[ $storage -eq 1 ]]
+then info analyse "$(red locli storage backend: database)"
+else info analyse "$(red locli storage backend: file)"
+fi
+
 progress "analyse" "args:  $(yellow $*)"
 while test $# -gt 0
 do case "$1" in
@@ -411,7 +420,10 @@ EOF
 
         local v0 v1 v2 v3 v4 v5 v6 v7 v8 v9 va vb vc vd ve vf vg vh vi vj vk vl vm vn vo
         v0=( $* )
-        v1=("${v0[@]/#logs/                 'unlog' --run-logs \"$adir\"/log-manifest.json ${analysis_allowed_loanys[*]/#/--ok-loany } }")
+        if [[ $storage -eq 1 ]]
+        then v1=("${v0[@]/#logs/            'unlog-db' --run-logs \"$adir\"/log-manifest-db.json }")
+        else v1=("${v0[@]/#logs/            'unlog' --run-logs \"$adir\"/log-manifest.json ${analysis_allowed_loanys[*]/#/--ok-loany } }")
+        fi
         v2=("${v1[@]/#read-context/         'read-meta-genesis'  --run-metafile    \"$dir\"/meta.json --shelley-genesis \"$dir\"/genesis-shelley.json }")
         v3=("${v2[@]/#write-context/        'write-meta-genesis' --run-metafile    \"$dir\"/meta.json --shelley-genesis \"$dir\"/genesis-shelley.json }")
         v4=("${v3[@]/#read-chain/           'read-chain'         --chain \"$adir\"/chain.json}")
@@ -452,6 +464,7 @@ EOF
                                              -e 'chain.json'            \
                                              -e 'hash-timeline.json'    \
                                              -e 'log-manifest.json'     \
+                                             -e 'log-manifest-db.json'  \
                                              -e 'mach-views.json'       \
                                              -e 'prof.json'             \
                                              -e 'tracefreq.json'
@@ -540,6 +553,12 @@ EOF
         ;;
 
     prepare | prep )
+        if [[ $storage -eq 1 ]]
+        then analyse prepare-db   "$@"
+        else analyse prepare-file "$@"
+        fi;;
+
+    prepare-file | prep-file )
         local usage="USAGE: wb analyse $op [[IDENT:]RUN-NAME=current].."
 
         local runspec=${1:-current}; if test $# != 0; then shift; fi
@@ -640,6 +659,90 @@ EOF
           grep -h 'TraceForgedBlock\|DownloadedHeader' $adir/logs-*.flt.json | sort > $ht_json
         fi;;
 
+    prepare-db | prep-db )
+        local usage="USAGE: wb analyse $op [--force] [[IDENT:]RUN-NAME=current].."
+
+        local remanifest_reasons=()
+        while test $# -gt 0; do
+          case "$1" in
+            --force ) remanifest_reasons+=("$(blue the --force was used)");;
+            * )       break;;
+          esac
+          shift
+        done
+
+        local runspec=${1:-current}; if test $# != 0; then shift; fi
+
+        ## Parse 'runspec' into either IDENT:RUN or RUN
+        local nrun=$(runspec_normalise $runspec)
+        local run=$(runspec_run $nrun)
+        local dir=$(run get "$run")
+        test -n "$dir" || fail "malformed run: $run"
+
+        progress "analyse(db)" "preparing run for analysis:  $(white $run)"
+
+        run trim "$run"
+        local adir=$dir/analysis
+        mkdir -p "$adir"/{cdf,png}
+
+        ## unless already done, filter logs to contain trace objects only
+        local logdirs=($(ls -d "$dir"/node-*/ 2>/dev/null))
+        local run_logs=$adir/log-manifest-db.json
+
+        test ${#logdirs[*]} -gt 0 ||
+            fail "Missing node-* subdirs in:  $dir"
+
+        if   test ! -f "$run_logs"
+        then remanifest_reasons+=("$(green missing $run_logs)")
+        fi
+
+        if test ${#remanifest_reasons[*]} = 0
+        then progress "analyse(db)" "log manifest up to date for raw logs"
+        else progress "analyse(db)" "assembling log manifest:  ${remanifest_reasons[*]}"
+             echo '{}' > $run_logs
+             # with useCabalRun we have to make sure the binary is built before launching off shell job spawns
+             local dummy=($(locli 2>/dev/null))
+             time {
+                 for d in "${logdirs[@]}"
+                 do throttle_shell_job_spawns
+                    local logfiles=($(ls --reverse -t "$d"stdout* "$d"node-[0-9]*.json \
+                                         2>/dev/null))
+                    if test -z "${logfiles[*]}"
+                    then msg "no logs in $d, skipping.."; fi
+                    local mach=$(basename "$d")
+                    local  out="$adir"/logs-$mach
+                    call_locli "silent" "serial" \
+                        prepare-db \
+                        --mach "$mach" \
+                        --db "$adir/logs-$mach.sqlite3" \
+                        ${logfiles[*]/#/--log } & 
+                    jq_fmutate "$run_logs" '
+                      .rlHostLogs["'"$mach"'"] =
+                        { hlRawLogfiles:    ["'"$(echo ${logfiles[*]} |
+                                                  sed 's/ /", "/')"'"]
+                        , hlRawLines:       0
+                        , hlRawTraceFreqs:  {}
+                        , hlLogs:           ["'"$adir/logs-$mach.sqlite3"'", null]
+                        , hlProfile:        []
+                        }
+                     | .rlFilterDate = ('$(if test -z "$without_datever_meta"
+                                           then echo -n now
+                                           else echo -n 0; fi)' | todate)
+                     '
+
+                    local ghc_rts_prof=$d/cardano-node.prof
+                    if test -f "$ghc_rts_prof"
+                    then progress "analyse(db) | profiling" "processing cardano-node.prof for $mach"
+                         ghc_rts_minusp_tojson "$ghc_rts_prof"           > "$out".flt.prof.json
+                         jq_fmutate "$run_logs" '
+                           .rlHostLogs["'"$mach"'"] += { hlProfile: $profile }
+                         ' --slurpfile profile "$out".flt.prof.json
+                    fi
+                 done
+                 wait
+             }
+        fi;;    
+
     trace-frequencies | trace-freq | freq | tf )
         local new_only= sargs=()
         while test $# -gt 0
@@ -686,10 +789,16 @@ EOF
 }
 
 call_locli() {
+    local silent=
+    while test $# -gt 0
+    do case "$1" in
+            silent ) silent='silent';;
+            * ) break;; esac; shift; done    
+    
     local rtsmode="${1:-hipar}"; shift
     local args=("$@")
 
-    echo "{ \"rtsmode\": \"$rtsmode\" }"
+    test "$silent" = "silent" || echo "{ \"rtsmode\": \"$rtsmode\" }"
     case "$rtsmode" in
         serial )locli_args+=(+RTS -N1 -A128M -RTS);;
         lomem ) locli_args+=(+RTS -N3 -A8M -RTS);;
@@ -698,7 +807,10 @@ call_locli() {
     esac
 
     verbose "analysis | locli" "$(with_color reset ${locli_args[@]}) $(colorise ${args[*]})"
-    time locli "${locli_args[@]}" "${args[@]}"
+    if test "$silent" = "silent"
+    then      locli "${locli_args[@]}" "${args[@]}"
+    else time locli "${locli_args[@]}" "${args[@]}"
+    fi
 }
 
 num_jobs="\j"